Idea
I might do a game in pygame using python. It would probably be similar to a tower defense game. Or I am going  to do something with arduino but I do not have an idea for what yet. Or, i might steal the idea you gave to the boys where you take a website and replace certain words in the website. I JUST DON'T KNOW!